delete comments and prints

+++ src/ReWork.java
@@ -1,6 +1,5 @@
 import java.util.ArrayList;
 
-import javafx.scene.layout.HBox;
 
 public class ReWork {
 
@@ -9,10 +8,9 @@ public class ReWork {
     public static int getBARework(ArrayList<ProjectStage> allStages){
         int bAReWork = 0;
         for (int i=0;i<allStages.size();i++){
-            if(!allStages.get(i).getIsNormal() && allStages.get(i).GetNewValue()<4){
+            if(!allStages.get(i).getIsNormal() && ((((int)allStages.get(i).GetNewValue())<4))){
 
                 bAReWork+=1;
-
             }
 
         }
@@ -20,9 +18,11 @@ public class ReWork {
     }
     public static int  getAARework(ArrayList<ProjectStage> allStages){
         int aAReWork = 0;
+        
         for (int i=0;i<allStages.size();i++){
-            if(!allStages.get(i).getIsNormal() && allStages.get(i).GetNewValue()>=4){
+            if(!allStages.get(i).getIsNormal() && (((int)allStages.get(i).GetNewValue())>=4)){
                 aAReWork+=1;
+
             }
         }
         return aAReWork;
